mod/requirements.txt
numpy
pandas
tqdm
Pillow
torchvision
scikit-learn
faiss-cpu
opencv-python
matplotlib
PyYAML
einops
